break up big file
