+++ tests/e2e/test_provider_oauth2_grafana.py
@@ -26,7 +26,8 @@ from authentik.providers.oauth2.models import (
     RedirectURIMatchingMode,
     ScopeMapping,
 )
-from tests.e2e.utils import SeleniumTestCase, retry
+from tests.decorators import retry
+from tests.selenium import SeleniumTestCase
 
 
 class TestProviderOAuth2OAuth(SeleniumTestCase):
