+++ tests/e2e/test_provider_radius.py
@@ -13,7 +13,8 @@ from authentik.flows.models import Flow
 from authentik.lib.generators import generate_id, generate_key
 from authentik.outposts.models import Outpost, OutpostConfig, OutpostType
 from authentik.providers.radius.models import RadiusProvider
-from tests.e2e.utils import E2ETestCase, retry
+from tests.decorators import retry
+from tests.live import E2ETestCase
 
 
 class TestProviderRadius(E2ETestCase):
