+++ tests/openid_conformance/base.py
@@ -7,8 +7,8 @@ from selenium.webdriver.support import expected_conditions as ec
 
 from authentik.blueprints.tests import apply_blueprint, reconcile_app
 from authentik.providers.oauth2.models import OAuth2Provider
-from tests.e2e.utils import SeleniumTestCase
 from tests.openid_conformance.conformance import Conformance
+from tests.selenium import SeleniumTestCase
 
 
 class TestOpenIDConformance(SeleniumTestCase):
